requirements.txt
pyaudiowpatch>=0.2.12.5
numpy>=1.24.0
scipy>=1.11.0

# VAD 模块依赖
torch>=2.0.0
torchaudio>=2.0.0
pyyaml>=6.0

# 记忆与检索模块依赖
chromadb>=0.4.0
sentence-transformers>=2.2.0
pydantic>=2.0
modelscope>=1.9.0

# 说话人识别模块依赖
pyannote.audio>=3.0.0
soundfile>=0.12.0
